remove blue from links, new structure for files, now each component has his own folder, bug fixed in loggin
