+++ src/assets/styles/containers/Container.scss
@@ -1,4 +1,4 @@
-.container {
+.containerMain {
   margin: 3.5em auto;
   width: 80%;
