+++ src/components/Admin/AdminAccount.scss
@@ -1,4 +1,4 @@
-@import '../components/App.scss';
+@import '../../assets/styles/components/App.scss';
 
 $font-family: 'Lato';
 $font-size: 16px;
