+++ src/components/Card/Card.scss
@@ -1,4 +1,4 @@
-@import '../components/App.scss';
+@import '../../assets/styles/components/App.scss';
 
 .card {
   background-color: var(--card);
