+++ src/components/CartProducts/CartProducts.jsx
@@ -1,8 +1,8 @@
 /* eslint-disable jsx-quotes */
 import React from 'react';
-import CartList from './CartList';
-import CartCards from './CartCards';
-import '../assets/styles/components/CartProducts.scss';
+import CartList from '../CartList/CartList';
+import CartCards from '../CartCards/CartCards';
+import './CartProducts.scss';
 
 const Products = ({ state }) => (
   <div>
