+++ src/components/Categories/Categories.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../assets/styles/components/Categories.scss';
+import './Categories.scss';
 
 const Categories = ({ children, title }) => (
   <>
