+++ src/components/Categories/Categories.scss
