+++ src/containers/Containers.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import '../assets/styles/containers/Container.scss';
 
 function Containers(props) {
-  return <div className="container">{props.children}</div>;
+  return <div className="containerMain">{props.children}</div>;
 }
 
 export default Containers;
