+++ src/routes/App.jsx
@@ -5,9 +5,9 @@ import '../assets/styles/components/Globales.scss';
 import { Link, BrowserRouter, Route, Switch } from 'react-router-dom';
 import Containers from '../containers/Containers';
 import Layout from '../components/Layout';
-import Categories from '../components/Categories';
-import ProductsContainer from '../components/ProductsContainer';
-import Card from '../components/Card';
+import Categories from '../components/Categories/Categories';
+import ProductsContainer from '../components/ProductsContainer/ProductsContainer';
+import Card from '../components/Card/Card';
 import Success from '../pages/Success';
 import UserAccount from '../pages/UserAccount';
 import AdminCMS from '../pages/AdminCMS';
